# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.AdminAuditLogResponse:
    properties:
      action:
        type: string
      actor_id:
        type: string
      actor_type:
        type: string
      correlation_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      ip_address:
        type: string
      new_value:
        additionalProperties: {}
        type: object
      old_value:
        additionalProperties: {}
        type: object
      request_id:
        type: string
      resource_id:
        type: string
      resource_name:
        type: string
      resource_type:
        type: string
      result:
        type: string
      tenant_id:
        type: string
      user_agent:
        type: string
    type: object
...
  dtos.AdminClientUsageResponse:
    properties:
      active_refresh_tokens:
        type: integer
      authorize_events_30d:
        type: integer
      client_id:
        type: string
      last_token_issued_at:
        type: string
      token_issue_events_30d:
        type: integer
      total_refresh_tokens:
        type: integer
    type: object
...
info:
  contact: {}
  description: GateForge IAM application API under /api/v1. The canonical public SDK
    contract (including root OIDC paths) lives at api/openapi.yaml in the monorepo.
  title: GateForge IAM API
  version: 0.1.0
paths:
...
  /admin/audit-logs:
    get:
      parameters:
      - description: Filter by tenant UUID
        in: query
        name: tenant_id
        type: string
      - description: Filter by action (exact or prefix with trailing dot)
        in: query
        name: action
        type: string
      - description: Filter by result (success, failure, denied)
        in: query
        name: result
        type: string
      - description: Filter by actor id
        in: query
        name: actor_id
        type: string
      - description: ISO8601 lower bound on created_at
        in: query
        name: from
        type: string
      - description: ISO8601 upper bound on created_at
        in: query
        name: to
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/dtos.AdminAuditLogResponse'
                type: array
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: List platform audit logs
      tags:
      - Admin
  /admin/clients:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by tenant UUID
        in: query
        name: tenant_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/dtos.AdminClientResponse'
                type: array
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: List OAuth clients (platform admin)
      tags:
      - Admin
    post:
      consumes:
      - application/json
      parameters:
      - description: OAuth client registration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.AdminCreateClientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.AdminCreateClientResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: Create an OAuth client (platform admin)
      tags:
      - Admin
  /admin/clients/{clientId}:
    delete:
      parameters:
      - description: Client record UUID
        in: path
        name: clientId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete an OAuth client (platform admin)
      tags:
      - Admin
    get:
      parameters:
      - description: Client record UUID
        in: path
        name: clientId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.AdminClientResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: Get an OAuth client by ID (platform admin)
      tags:
      - Admin
    patch:
      consumes:
      - application/json
      parameters:
      - description: Client record UUID
        in: path
        name: clientId
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.AdminUpdateClientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.AdminClientResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: Update an OAuth client (platform admin)
      tags:
      - Admin
  /admin/clients/{clientId}/usage:
    get:
      parameters:
      - description: Client record UUID
        in: path
        name: clientId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.AdminClientUsageResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: Get OAuth client usage metrics (platform admin)
      tags:
      - Admin
  /admin/login-history:
    get:
      parameters:
      - description: Filter by tenant UUID
        in: query
        name: tenant_id
        type: string
      - description: Filter by result (success, failure, denied)
        in: query
        name: result
        type: string
      - description: Filter by actor id
        in: query
        name: actor_id
        type: string
      - description: ISO8601 lower bound on created_at
        in: query
        name: from
        type: string
      - description: ISO8601 upper bound on created_at
        in: query
        name: to
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/dtos.AdminAuditLogResponse'
                type: array
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: List login history events (platform admin)
      tags:
      - Admin
  /admin/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.AdminStatsResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: Platform admin dashboard stats
      tags:
      - Admin
  /admin/tenants:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/dtos.AdminTenantResponse'
                type: array
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: List tenants (platform admin)
      tags:
      - Admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant name and optional domain
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.AdminCreateTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.AdminTenantResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: Create a tenant (platform admin)
      tags:
      - Admin
  /admin/tenants/{tenantId}:
    delete:
      parameters:
      - description: Tenant UUID
        in: path
        name: tenantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete a tenant (platform admin)
      tags:
      - Admin
    get:
      parameters:
      - description: Tenant UUID
        in: path
        name: tenantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.AdminTenantResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: Get a tenant by ID (platform admin)
      tags:
      - Admin
    patch:
      consumes:
      - application/json
      parameters:
      - description: Tenant UUID
        in: path
        name: tenantId
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.AdminUpdateTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.AdminTenantResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: Update a tenant (platform admin)
      tags:
      - Admin
  /admin/tenants/{tenantId}/identity-providers:
    get:
      parameters:
      - description: Tenant UUID
        in: path
        name: tenantId
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/dtos.AdminIdentityProviderResponse'
                type: array
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: List identity providers for a tenant
      tags:
      - Admin
  /admin/tenants/{tenantId}/identity-providers/{provider}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Tenant UUID
        in: path
        name: tenantId
        required: true
        type: string
      - description: Provider id (e.g. google)
        in: path
        name: provider
        required: true
        type: string
      - description: OAuth credentials and enabled flag
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.PatchIdentityProviderRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Configure an upstream identity provider for a tenant (platform admin)
      tags:
      - Admin
  /admin/tenants/{tenantId}/members:
    get:
      parameters:
      - description: Tenant UUID
        in: path
        name: tenantId
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/dtos.AdminTenantMemberResponse'
                type: array
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: List members of a tenant (platform admin)
      tags:
      - Admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant UUID
        in: path
        name: tenantId
        required: true
        type: string
      - description: User email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.AdminAddMemberRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Add an existing user to a tenant
      tags:
      - Admin
  /admin/tenants/{tenantId}/members/{userId}:
    delete:
      parameters:
      - description: Tenant UUID
        in: path
        name: tenantId
        required: true
        type: string
      - description: User UUID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Remove a user from a tenant
      tags:
      - Admin
  /admin/users:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by tenant UUID
        in: query
        name: tenant_id
        type: string
      - description: Search email or name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/dtos.AdminUserResponse'
                type: array
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: List users (platform admin)
      tags:
      - Admin
  /admin/users/{userId}:
    get:
      parameters:
      - description: User UUID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.AdminUserDetailResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: Get user details (platform admin)
      tags:
      - Admin
  /admin/users/{userId}/disable:
    post:
      parameters:
      - description: User UUID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Disable a user account (platform admin)
      tags:
      - Admin
  /admin/users/{userId}/force-logout:
    post:
      parameters:
      - description: User UUID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Force logout a user (platform admin)
      tags:
      - Admin
  /admin/users/{userId}/reset-mfa:
    post:
      parameters:
      - description: User UUID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Reset MFA (TOTP + recovery codes) for a user (platform admin)
      tags:
      - Admin
  /admin/users/{userId}/reset-passkey:
    post:
      parameters:
      - description: User UUID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Reset all passkeys for a user (platform admin)
      tags:
      - Admin
  /authorize:
    get:
      consumes:
      - application/json
      parameters:
      - description: Session cookie from browser login (iam_session)
        in: header
        name: Cookie
        type: string
      - description: State parameter
        in: query
        name: state
        type: string
      - description: Response type (code)
        in: query
        name: response_type
        type: string
      - description: Client ID
        in: query
        name: client_id
        type: string
      - description: Redirect URI
        in: query
        name: redirect_uri
        type: string
      - description: Scope
        in: query
        name: scope
        type: string
      - description: Nonce
        in: query
        name: nonce
        type: string
      - description: Code challenge
        in: query
        name: code_challenge
        type: string
      - description: Code challenge method
        in: query
        name: code_challenge_method
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to client or login
          schema:
            type: string
      summary: Authorize endpoint (authorization code + PKCE)
      tags:
      - OIDC
  /federation/providers:
    get:
      parameters:
      - description: Tenant UUID (defaults to DEFAULT_TENANT_ID)
        in: query
        name: tenant_id
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/dtos.PublicFederationProviderResponse'
                type: array
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      summary: List enabled federation sign-in methods for a tenant
      tags:
      - Auth
  /forgot-password:
    post:
      consumes:
      - application/json
      - application/json
      description: Always returns 200 so callers cannot tell whether the email exists
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.RefreshTokenRequest'
      - description: Email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.ForgotPasswordRequest'
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      summary: Request a password reset
      tags:
      - Auth
      - Auth
...
  /internal/tenants/{tenantId}/identity-providers/{provider}:
    patch:
      parameters:
      - description: Tenant UUID
        in: path
        name: tenantId
        required: true
        type: string
      - description: Provider id (e.g. google)
        in: path
        name: provider
        required: true
        type: string
      - description: Admin API key
        in: header
        name: X-Admin-API-Key
        required: true
        type: string
      - description: OAuth credentials and enabled flag
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.PatchIdentityProviderRequest'
      responses:
        "204":
          description: No Content
      summary: Configure an upstream identity provider for a tenant
      tags:
      - internal
  /introspect:
    post:
      consumes:
      - application/json
      - application/x-www-form-urlencoded
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.TokenIntrospectionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Token introspection endpoint (RFC 7662)
      tags:
      - OIDC
  /login:
    post:
      consumes:
      - application/json
      description: Sets the same iam_session cookie as OIDC so /authorize recognizes
        the browser without a second login. If the user has MFA (TOTP) enabled, `data`
        is MFALoginChallengeResponse (mfa_ticket) instead of LoginResponse; complete
        login with POST /mfa/challenge/verify. When the user belongs to multiple tenants
        and no tenant context is provided, `data` is TenantSelectionResponse.
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'When MFA is off: LoginResponse. When MFA is on: same envelope
            but `data` is MFALoginChallengeResponse (see top-level description).'
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.LoginResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      summary: Login (dashboard/API) with email and password
      tags:
      - Auth
  /login/session:
    post:
      description: Reads the browser iam_session cookie set by federation callback
        and returns JWT tokens for the SPA dashboard. Used by GET /login/federation/complete.
      produces:
      - application/json
      responses:
        "200":
          description: When tenant is resolved. When multiple tenants apply, `data`
            is TenantSelectionResponse.
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.LoginResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      summary: Exchange iam_session cookie for API tokens (after federated sign-in)
      tags:
      - Auth
  /logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      security:
      - BearerAuth: []
      summary: Log out everywhere for the current user (SSO session + all refresh
        tokens)
      tags:
      - Auth
...
schemes:
- http
- https
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Bearer Token Authentication. Use "Bearer {token}" as the value.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"